consistent use of "int" rather than "long"
git-svn-id: https://src.springframework.org/svn/spring-maintenance/trunk@3701 fd5a2b45-1f63-4059-99e9-3c7cb7fd75c8

+++ samples/petclinic/src/org/springframework/samples/petclinic/jdbc/AbstractJdbcClinic.java
@@ -107,7 +107,7 @@ abstract public class AbstractJdbcClinic extends JdbcDaoSupport implements Clini
 			List vetSpecialtiesIds = this.vetSpecialtiesQuery.execute(vet.getId());
 			Iterator vsi = vetSpecialtiesIds.iterator();
 			while (vsi.hasNext()) {
-				long specialtyId = ((Integer) vsi.next()).longValue();
+				int specialtyId = ((Integer) vsi.next()).intValue();
 				Specialty specialty = (Specialty) EntityUtils.getById(specialties, Specialty.class, specialtyId);
 				vet.addSpecialty(specialty);
 			}
